Jarvis AI/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==1.3.5
SpeechRecognition>=3.14.0
pyttsx3==2.90
PyAudio==0.2.14
httpx==0.25.1
wikipedia==1.4.0
requests==2.31.0
python-multipart==0.0.9
Pillow>=11.0.0
google-generativeai>=0.8.0
beautifulsoup4==4.12.3
sse-starlette==2.0.0
